+++ packages/anki/src/index.ts
@@ -1,5 +1,6 @@
 export * from "./ankiconnect.ts";
 export * from "./abort.ts";
+export * from "./launch.ts";
 export {
   classifyConnectError,
   ANKICONNECT_ADDON_CODE,